+++ spec/lib/journey_spec.rb
@@ -3,7 +3,7 @@ require_relative "../../app/lib/journey"
 
 RSpec.describe Journey do
   let(:store) { {} }
-  let(:form_context) { FormContext.new(store) }
+  let(:form_context) { Flow::FormContext.new(store) }
   let(:step_factory) { StepFactory.new(form:) }
   let(:journey) { described_class.new(form_context:, step_factory:) }
